+++ modules/smithplates-http-service-renderer/src/main/scala/com/jacoby6000/smithplates/http/service/renderer/HttpServiceCodegenRenderer.scala
@@ -98,7 +98,14 @@ object HttpServiceCodegenRenderer {
         templatePath: String,
         view: TemplateView[S, M]
     ): CodegenValidated[String] =
-      renderTemplateAttributes(settings, templatePath, Map("ctx" -> view))
+      renderTemplateAttributes(
+        settings,
+        templatePath,
+        Map(
+          "ctx"         -> view,
+          "httpLibrary" -> settings.defaultFrameworkKey
+        )
+      )
 
     def renderTemplateAttributes(
         settings: HttpServiceCodegenSettings,